# Use reduction_count=2 in sha256 example.