@mpenciak 2023-08-29
I originally implemented it with `rc = 10` when the circuit was something like `(eq (sha256) (0xf5...))` and it took 7 reductions. When I moved the equality check into the coprocessor I didn't change the `rc` count to minimize... I'm not sure how it got to 100 though.