@porcuquine 2023-08-29
It looks like it happened here: https://github.com/lurk-lab/lurk-rs/commit/34a3e657c1857bf5eb4cadbb33bf438f0e2b18d4#diff-395091e3dca52327c5b3bb679cb6ad0344a2add805575c70f9f4507d94d20f17R29. @samuelburnham Do you have any memory of why you changed it? If this was the consequence of blind merging, we might have made other unintended and unnoticed changes, so it might be worth looking more carefully now…