@winston-h-zhang 2023-08-29
I think this may have been me at some point, since I was playing around with public parameters. Sam probably rebased on my changes.