@samuelburnham 2023-08-29
Yes I just rebased on top of dev at the time. I'm not seeing any other changes in #620  that would have adversely affected the examples/benchmarks but maybe others can take a look.